% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Data plotting}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the corresponding state}

\item{year}{An integer representing the year for which the plot is made}
}
\value{
This returns a map with the accidents plotted for the state in the input
}
\description{
This function plots the accidents for the state provided in the input on the map. The function throws an error if the state number in the input argument does not exist in the fars data.If there are no accidents reported for the state of interest, a message is displayed.
}
\examples{
fars_map_state(3,2013)

}
