% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename of the data}
}
\value{
This function returns a data frame
}
\description{
This function reads data in compressed .csv format and creates a data frame. It throws an error if the filename in the argument does not exist.
}
\examples{
\dontrun{fars_read("Data.csv.bz2")}

}
