% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract a subset of data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years}
}
\value{
This function returns a list of data frames for each year in the input and none if there is no data available for a year. If there is no data available for a year, the function throws a warning.
}
\description{
This function creates a data list with the month and year columns for each year in the input
}
\examples{
fars_read_years(list(2013,2014,2015))

}
