% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Data operations}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years}
}
\value{
This returns a data frame with each column for each year in the input argument and the corresponding number of accidents for that year.
}
\description{
This function groups data by year and month and summarizes the total number of accidents for each year
}
\examples{
\dontrun{fars_summarize_years(list(2013,2014,2015))}

}
