% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filenames}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer giving the year}
}
\value{
This function returns a string with the year
}
\description{
This function creates filenames with extensions with the year input
}
\examples{
\dontrun{make_filename(2013)}

}
